splitted the auth module
