+++ services/utility/fishapp_rest_client/src/main/java/no/fishapp/util/restClient/AuthBaseClientInterface.java
@@ -1,4 +1,4 @@
-package no.fishapp.util.restClient.auth;
+package no.fishapp.util.restClient;
 
 
 public interface AuthBaseClientInterface {
