--- services/utility/src/main/java/no/fishapp/util/restClient/exceptionHandlers/RestClientHttpException.java
@@ -1,17 +1,0 @@
-package no.fishapp.util.restClient.exceptionHandlers;
-
-import javax.ws.rs.core.Response;
-
-public class RestClientHttpException extends Exception {
-
-    int httpStatusCode;
-
-    public int getHttpStatusCode() {
-        return httpStatusCode;
-    }
-
-    public RestClientHttpException(int httpStatusCode) {
-        this.httpStatusCode = httpStatusCode;
-
-    }
-}
